requirements.txt
click
colorama
Flask
Flask-Login==0.6.3
importlib-metadata
itsdangerous
Jinja2
MarkupSafe
Werkzeug
zipp
python-dotenv
